Fix `showpage` and `kidsonly` when requeted page is a subpage + show displaytitle
Change 1: Pass requested root page to the HierarchyCreator
This is done so that it can know which page is the actual root instead of just going up until
there are no more parent pages.

This will force "one root page rule", which is already enforced anyway

Change 2: Show displaytitle of the page if available
Added a new "pathStyle" => displaytitle, which will try to get the displaytitle
of the page to display. If not found, it will fall-back to `subpage` path style

Warning: this is now set as a default